sdbfs: relicense lib to LGPL v2.1
The userspace and kernel binaries are kept as GPL v2 but the library has
been relicensed as LGPL v2.1 in order to allow it to be linked to
proprietary software.

The permission to relicense the files has been approved by CERN in 2018
and then in 2021 (and also by its main author).